legacy/dining_dashboard/requirements.txt
streamlit
pandas
matplotlib
PyYAML
streamlit-authenticator>=0.4.1
bcrypt
